fix: Default exports and declarations for Buffer and Events module
