+++ node/_events.js
@@ -1,5 +1,3 @@
-// @deno-types="./events.d.ts"
-
 // Copyright Joyent, Inc. and other Node contributors.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a
