+++ node/buffer.ts
@@ -1,5 +1,2 @@
 // @deno-types="./_buffer.d.ts"
-import Buffer from "./_buffer.js";
-export { Buffer } from "./_buffer.js";
-
-export default Buffer;
+export { Blob, Buffer, default } from "./_buffer.js";
