+++ node/events.ts
@@ -1,0 +1,11 @@
+// @deno-types="./_events.d.ts"
+export {
+  captureRejectionSymbol,
+  default,
+  defaultMaxListeners,
+  errorMonitor,
+  EventEmitter,
+  getEventListeners,
+  on,
+  once,
+} from "./_events.js";
